+++ VideoSample/src/main/java/com/hyh/video/lib/MediaEventListener.java
@@ -12,15 +12,15 @@ public interface MediaEventListener {
 
     void onPreparing();
 
-    void onPrepared(int duration);
+    void onPrepared(long duration);
 
-    void onStart(int currentPosition, int duration);
+    void onStart(long currentPosition, long duration);
 
-    void onPlaying(int currentPosition, int duration);
+    void onPlaying(long currentPosition, long duration);
 
-    void onPause(int currentPosition, int duration);
+    void onPause(long currentPosition, long duration);
 
-    void onStop(int currentPosition, int duration);
+    void onStop(long currentPosition, long duration);
 
     void onBufferingStart();
 
@@ -28,7 +28,7 @@ public interface MediaEventListener {
 
     void onBufferingUpdate(int progress);
 
-    void onSeekStart(int seekMilliSeconds, int seekProgress);
+    void onSeekStart(long seekMilliSeconds, int seekProgress);
 
     void onSeekEnd();
 
@@ -38,5 +38,5 @@ public interface MediaEventListener {
 
     void onCompletion();
 
-    void onRelease(int currentPosition, int duration);
+    void onRelease(long currentPosition, long duration);
 }
